Regarding issue #239
Added a new static method to `Runtime` called `getLoadedLibraries()` which allows consumers to query all loaded libraries at any given point in time. This method returns a `List<LoadedLibraryData>` which is a simple data class that holds information about a loaded library such as names, search paths and successful paths. `getLoadedLibraries()` only returns *loaded* libraries, that is, libraries that have been loaded into memory successfully, ie a successful call to `dlopen()` was made. A library that has not yet loaded (because of lazy loading) will not appear, and a library that has been *unloaded* because it was GC'd will also not appear.